+++ libs/product/src/resolvers/public_api.ts
@@ -1,0 +1,1 @@
+export { DaffProductPageServerSideResolver } from './product-page-server-side/product-page-server-side.resolver';
